requirements.txt
beautifulsoup4
langchain
google-generativeai
streamlit
langchain-google-genai
python-dotenv
chromadb
langchain_community
transformers
